feat(council): modularize orchestrator into stages package (#28)
- Refactored monolithic council.py into llm_council.stages package
- Implemented Stage 1 (Ideation), Stage 2 (Review), and Stage 3 (Synthesis)
- Maintained 100% backward compatibility via Facade pattern and __getattr__ delegation
- Implemented Patch-Aware Gateway to support legacy test mocks
- Stabilized progress reporting bridge for MCP/CLI parity
- Resolved quality gate regressions (Ruff E701/Mypy)
- Verified with full regression suite (37/37 tests passed)
