+++ docs/progress_reports/progress_report_04102026-1655.md
@@ -1,0 +1,29 @@
+# Session Progress Report - April 10, 2026
+
+## Objective
+Finalize stabilization, authentication, and interface polish for the LLM Council MCP server and CLI.
+
+## Key Achievements
+*   **Resolved Runtime Crashes**: Fixed the `unhashable type: dict` error in the MCP result formatter, allowing multi-model rankings to display correctly in Claude Desktop.
+*   **Dependency & Import Stabilization**: Corrected multiple broken internal import paths (`unified_config`, `tier_contract`, `verification`) in `mcp_server.py` that were causing server startup failures.
+*   **Authentication Fix**: Updated the server to use the `get_api_key()` resolver instead of raw config access, allowing successful API key retrieval from the Windows Credential Manager.
+*   **Terminology & UI Polish**: 
+    *   Renamed the Devil's Advocate output to **"ADVERSARIAL CRITIQUE"** to distinguish it from Stage 2 Constructive Dissent.
+    *   Implemented header scrubbing to remove redundant model-injected prefixes (e.g., "**claude-3-haiku**:") from the reports.
+    *   Reordered MCP tools for better UX (Health Check first, monolithic tool last).
+*   **E2E Verification**: Successfully verified the 3-stage pipeline and adversarial audit via the `query.py` CLI.
+*   **Architectural Unification**: Refactored the monolithic `run_full_council` to delegate to modular stages (`run_stage1/2/3`), eliminating 250+ lines of duplicate logic and ensuring parity between CLI and MCP server behavior.
+
+## 📂 Modified Files
+*   `src/llm_council/mcp_server.py`
+*   `src/llm_council/council.py`
+*   `query.py` (project root)
+
+## Current State
+*   **MCP Server**: Stable and fully functional in Claude Desktop.
+*   **CLI**: `query.py` and `llm-council` commands verified on Windows.
+*   **Auth**: Windows Keychain (Credential Manager) integration is active and verified.
+
+## Next Steps
+*   **Monitoring**: Continuous monitoring of Claude Desktop logs for any latent "unclear" verdict edge cases.
+*   **Documentation**: Update the README to reflect the 3-stage tool sequence requirement for MCP users.
