+++ src/llm_council/constants.py
@@ -1,0 +1,20 @@
+"""Centralized constants for LLM Council orchestration."""
+
+from llm_council.gateway_adapter import (
+    STATUS_OK,
+    STATUS_TIMEOUT,
+    STATUS_ERROR,
+    STATUS_RATE_LIMITED,
+)
+
+# ADR-012: Tiered Timeout Strategy Constants
+TIMEOUT_PER_MODEL_SOFT = 15.0
+TIMEOUT_PER_MODEL_HARD = 25.0
+TIMEOUT_SYNTHESIS_TRIGGER = 40.0
+TIMEOUT_RESPONSE_DEADLINE = 50.0
+
+# ADR-012: Model Status Types
+MODEL_STATUS_OK = STATUS_OK
+MODEL_STATUS_TIMEOUT = STATUS_TIMEOUT
+MODEL_STATUS_ERROR = STATUS_ERROR
+MODEL_STATUS_RATE_LIMITED = STATUS_RATE_LIMITED
