+++ src/llm_council/mcp_server.py
@@ -1,227 +1,108 @@
-"""LLM Council MCP Server - consult multiple LLMs and get synthesized guidance.
-
-Implements ADR-012: MCP Server Reliability and Long-Running Operation Handling
-- Progress notifications during council execution
-- Health check tool
-- Confidence levels (quick/balanced/high/reasoning)
-- Structured results with per-model status
-- Tiered timeouts with fallback synthesis
-- Partial results on timeout
-- Tier-Sovereign timeout configuration (2025-12-19)
-"""
-
-import json
-import time
-import asyncio
 import os
 import sys
-from typing import List, Optional, Any, Dict, Callable
+import time
+import json
+import uuid
+import asyncio
+import logging
+import random
+from typing import Dict, List, Any, Optional, Callable, Tuple, Awaitable
 
-# Ensure src is in sys.path for robust imports (ADR-032 path fix)
-BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
-if BASE_DIR not in sys.path:
-    sys.path.insert(0, BASE_DIR)
-
+# MCP SDK imports
 from mcp.server.fastmcp import FastMCP, Context
 
+# Inject src directory into sys.path to ensure local imports work regardless of execution context
+# This is critical for Claude Desktop/uv which may not have the package installed in editable mode
+src_path = os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))
+if src_path not in sys.path:
+    sys.path.insert(0, src_path)
+
+# Local imports
 from llm_council.council import (
+    run_stage1, 
+    run_stage2, 
+    run_stage3, 
     run_council_with_fallback,
-    run_stage1,
-    run_stage2,
-    run_stage3,
-    TIMEOUT_SYNTHESIS_TRIGGER,
+    COUNCIL_MODELS,
+    CHAIRMAN_MODEL
 )
 from llm_council.tier_contract import create_tier_contract
-from llm_council.verdict import VerdictType
+from llm_council.openrouter import STATUS_OK, STATUS_ERROR
+from llm_council.gateway_adapter import query_model_with_status
+from llm_council.unified_config import get_config, get_key_source, get_api_key
 from llm_council.session_store import (
-    create_session,
-    load_session,
-    save_session,
-    close_session,
-    purge_expired_sessions,
+    create_session, 
+    load_session, 
+    save_session, 
+    close_session, 
+    purge_expired_sessions
 )
 from llm_council.verification.api import run_verification, VerifyRequest
-from llm_council.verification.context import InvalidSnapshotError
 from llm_council.verification.formatting import format_verification_result
-from llm_council.verification.transcript import (
-    create_transcript_store,
-    TranscriptNotFoundError,
-    TranscriptIntegrityError,
-)
+from llm_council.verification.context import InvalidSnapshotError
+from llm_council.verification.transcript import create_transcript_store
 
-# ADR-032: Migrated to unified_config
-from llm_council.unified_config import get_config, get_api_key
-from llm_council.openrouter import query_model_with_status, STATUS_OK
+# Initialize FastMCP server
+mcp = FastMCP("llm-council")
 
+# --- Helper Functions ---
 
-def _get_council_models() -> list:
-    """Get council models from unified config."""
-    return get_config().council.models
-
-
-def _get_chairman_model() -> str:
-    """Get chairman model from unified config."""
-    return get_config().council.chairman
+# ADR-012: Predefined confidence tiers for easy client consumption
+CONFIDENCE_CONFIGS = {
+    "quick": {"models": 2, "description": "Fast and cheap"},
+    "balanced": {"models": 3, "description": "Balanced cost and quality"},
+    "high": {"models": None, "description": "Comprehensive review (all models)"},
+}
 
 
-def _get_openrouter_api_key() -> str:
-    """Get OpenRouter API key via ADR-013 resolution chain."""
-    return get_api_key("openrouter") or ""
+def _get_openrouter_api_key() -> Optional[str]:
+    """Lazy helper to resolve API key via standard priority chain."""
+    return get_api_key("openrouter")
 
-
-def _get_tier_model_pools() -> dict:
-    """Get tier model pools from unified config."""
+def _get_tier_model_pools() -> Dict[str, List[str]]:
+    """Lazy helper to get tier model pools from config."""
     config = get_config()
-    return config.tiers.pools
+    return {name: pool.models for name, pool in config.tiers.pools.items()}
 
-
-def _get_tier_timeout(tier: str) -> dict:
-    """Get tier timeout config from unified config."""
+def _get_tier_timeout(tier: str) -> Dict[str, int]:
+    """Lazy helper to get tier timeouts from config."""
     config = get_config()
-    timeouts = config.timeouts
-    return {
-        "total": timeouts.get_timeout(tier, "total") // 1000,  # Convert ms to seconds
-        "per_model": timeouts.get_timeout(tier, "per_model") // 1000,
-    }
+    pool = config.tiers.pools.get(tier)
+    if pool:
+        return {"per_model": pool.timeout_seconds, "total": pool.timeout_seconds * 2}
+    return {"per_model": 90, "total": 180}
 
-
-def _get_key_source() -> str:
-    """Determine the source of the API key."""
-    import os
-
-    if os.environ.get("OPENROUTER_API_KEY"):
-        return "environment"
-    # Could add keychain detection here
-    return "unknown"
-
-
-# Module-level function for backwards compatibility with tests
-def get_key_source() -> str:
-    """Public function wrapper for backwards compatibility."""
-    return _get_key_source()
-
-
-# Module-level aliases for backwards compatibility
-COUNCIL_MODELS = _get_council_models()
-CHAIRMAN_MODEL = _get_chairman_model()
-OPENROUTER_API_KEY = _get_openrouter_api_key()
-TIER_MODEL_POOLS = _get_tier_model_pools()
-
-
-mcp = FastMCP("LLM Council")
-
-
-def _build_confidence_configs() -> dict:
+def _get_progress_callback(ctx: Context) -> Optional[Callable]:
+    """Create a progress callback that bridges to MCP context.
+    
+    Supports both modern `ctx.info()` and legacy `ctx.report_progress()` for compatibility.
     """
-    Build confidence configs dynamically from tier timeout settings.
-
-    This allows environment variable overrides per ADR-012 Section 5.
-    """
-    return {
-        "quick": {
-            "models": 2,
-            **_get_tier_timeout("quick"),
-            "description": "Fast response (~20-30s)",
-        },
-        "balanced": {
-            "models": 3,
-            **_get_tier_timeout("balanced"),
-            "description": "Balanced response (~45-60s)",
-        },
-        "high": {
-            "models": None,
-            **_get_tier_timeout("high"),
-            "description": "Full council deliberation (~90s)",
-        },
-        "reasoning": {
-            "models": None,
-            **_get_tier_timeout("reasoning"),
-            "description": "Deep reasoning models (~3-5min)",
-        },
-    }
-
-
-# Build configs at import time (can be refreshed if needed)
-CONFIDENCE_CONFIGS = _build_confidence_configs()
-
-
-def _get_progress_callback(ctx: Optional[Context]) -> Optional[Callable]:
-    """
-    Creates a progress callback bridge for a given MCP context.
-    Implements ADR-012 fire-and-forget logic in prod, and synchronous await in tests.
-    """
-    if not ctx:
+    if ctx is None:
         return None
 
-    import os, sys
-    in_test = os.environ.get("PYTEST_CURRENT_TEST") or "pytest" in sys.modules
-
     async def on_progress(step: int, total: int, message: str):
-        async def _send():
-            try:
-                # FastMCP ctx.report_progress(step, total)
-                await asyncio.wait_for(ctx.report_progress(step, total), timeout=0.5)
-            except Exception:
-                pass
-            try:
-                # ADR-012: ctx.info() pushes messages into Claude's "Thinking" block
-                await asyncio.wait_for(ctx.info(message), timeout=0.5)
-            except Exception:
-                pass
+        """Bridge on_progress to MCP Context."""
+        import inspect
 
-        if in_test:
-            # We must await in tests to avoid race conditions and test failures
-            await _send()
-        else:
-            # Fire-and-forget in prod to prevent stdio pipe blocking/deadlocks
-            asyncio.create_task(_send())
+        # 1. Modern FastMCP interface (ctx.info for status messages)
+        try:
+            if hasattr(ctx, "info"):
+                await ctx.info(message)
+        except Exception:
+            pass
+
+        # 2. Legacy / Test Mock interface (ctx.report_progress for numeric status)
+        try:
+            if hasattr(ctx, "report_progress"):
+                # Following standard MCP Context.report_progress(completed, total)
+                res = ctx.report_progress(step, total)
+                
+                if inspect.isawaitable(res):
+                    await res
+        except Exception:
+            pass
 
     return on_progress
-
-
-@mcp.tool()
-async def consult_council(
-    query: str,
-    confidence: str = "balanced",
-    include_details: bool = False,
-    verdict_type: str = "synthesis",
-    include_dissent: bool = False,
-    adversarial_mode: bool = False,
-    ctx: Context = None,
-) -> str:
-    """
-    DEPRECATED: Use start_council -> council_review -> council_synthesize instead.
-    This monolithic tool is DISABLED for MCP use to ensure real-time progress visibility.
-    """
-    import os, sys
-    in_test = os.environ.get("PYTEST_CURRENT_TEST") or "pytest" in sys.modules
-    
-    if ctx is not None and not in_test:
-        return (
-            "ERROR: consult_council is disabled. You must use the three-stage flow to get visibility into deliberation stages:\n"
-            "1. start_council(query=..., confidence=...)\n"
-            "2. council_review(session_id=...)\n"
-            "3. council_synthesize(session_id=...)\n"
-            "Please use the split tools for this query."
-        )
-
-    # Fallback for direct Python calls (tests / CLI)
-    from llm_council.verdict import VerdictType
-    v_type = VerdictType(verdict_type) if verdict_type != "synthesis" else None
-
-    # Get standard progress bridge (ADR-012)
-    on_progress = _get_progress_callback(ctx)
-
-    council_result = await run_council_with_fallback(
-        query,
-        tier_contract=create_tier_contract(confidence),
-        verdict_type=v_type,
-        include_dissent=include_dissent,
-        adversarial_mode=adversarial_mode,
-        on_progress=on_progress,
-    )
-
-    return _format_council_result(council_result, include_details=include_details)
 
 
 def _format_council_result(council_result: Dict[str, Any], include_details: bool = False) -> str:
@@ -311,14 +192,18 @@ def _format_council_result(council_result: Dict[str, Any], include_details: bool
     # ADR-DA: Display Devil's Advocate Critique if available (Stage 1B)
     dissent_report = metadata.get("dissent_report")
     if dissent_report:
+        import re
+        # Strip redundant model prefixes or "Dissenting Report:" if present
+        cleaned_report = re.sub(r"^\*\*.*?\*\*:\s*", "", dissent_report)
+        cleaned_report = re.sub(r"^Dissenting Report:\s*", "", cleaned_report, flags=re.IGNORECASE)
+
         result += "\n" + "!" * 40 + "\n"
-        result += "### DEVIL'S ADVOCATE - DISSENTING REPORT\n"
+        result += "### DEVIL'S ADVOCATE - ADVERSARIAL CRITIQUE\n"
         result += "-" * 40 + "\n"
-        result += dissent_report + "\n"
+        result += f"Dissenting Report:\n{cleaned_report}\n"
         result += "!" * 40 + "\n"
 
     # ADR-CD: Display Constructive Dissent if available (Minority Opinion from Stage 2)
-    # Note: query.py calls this 'dissent', mcp_server metadata has it as 'dissent'
     minority_opinion = metadata.get("dissent")
     if minority_opinion:
         result += "\n" + "." * 40 + "\n"
@@ -376,11 +261,15 @@ def _format_council_result(council_result: Dict[str, Any], include_details: bool
         result += "\n#### Stage 1: Individual Opinions\n"
         # Recover label mappings to show models by name
         label_to_model = metadata.get("label_to_model", {})
-        model_to_label = {v: k for k, v in label_to_model.items()}
+        model_to_label = {}
+        for k, v in label_to_model.items():
+            if isinstance(v, dict) and "model" in v:
+                model_to_label[v["model"]] = k
+            else:
+                model_to_label[v] = k
 
         # Note: stage1_results is often in the top level council_result for CLI,
         # but in MCP metadata it's sometimes stored differently.
-        # We check both locations, and fall back to model_responses as a last resort.
         stage1_responses = metadata.get("stage1_results") or council_result.get("stage1_results")
 
         if not stage1_responses:
@@ -417,6 +306,70 @@ def _format_council_result(council_result: Dict[str, Any], include_details: bool
     return result
 
 
+# --- MCP Tools ---
+
+@mcp.tool()
+async def council_health_check() -> str:
+    """
+    Check LLM Council health before expensive operations (ADR-012).
+
+    Returns status of API connectivity, configured models, and estimated response time.
+    Use this to verify the council is working before calling start_council.
+    """
+    from importlib.metadata import version as pkg_version
+
+    try:
+        council_version = pkg_version("llm-council-core")
+    except Exception:
+        council_version = "unknown"
+
+    checks = {
+        "version": council_version,
+        "api_key_configured": bool(_get_openrouter_api_key()),
+        "key_source": get_key_source(),
+        "council_size": len(COUNCIL_MODELS),
+        "chairman_model": CHAIRMAN_MODEL,
+        "models": COUNCIL_MODELS,
+        "estimated_duration": {
+            "quick": "~20-30 seconds (fastest models)",
+            "balanced": "~45-60 seconds (most models)",
+            "high": f"~60-90 seconds (all {len(COUNCIL_MODELS)} models)",
+        },
+    }
+
+    if checks["api_key_configured"]:
+        try:
+            start = time.time()
+            response = await query_model_with_status(
+                "google/gemini-2.0-flash-001",
+                [{"role": "user", "content": "ping"}],
+                timeout=10.0,
+            )
+            latency_ms = int((time.time() - start) * 1000)
+
+            checks["api_connectivity"] = {
+                "status": response["status"],
+                "latency_ms": latency_ms,
+                "test_model": "google/gemini-2.0-flash-001",
+            }
+
+            if response["status"] == STATUS_OK:
+                checks["ready"] = True
+                checks["message"] = "Council is ready. Use start_council to ask questions."
+            else:
+                checks["ready"] = False
+                checks["message"] = f"API connectivity issue: {response.get('error', 'Unknown error')}"
+        except Exception as e:
+            checks["api_connectivity"] = {"status": "error", "error": str(e)}
+            checks["ready"] = False
+            checks["message"] = f"Health check failed: {e}"
+    else:
+        checks["ready"] = False
+        checks["message"] = "OPENROUTER_API_KEY not configured. Set it in environment or .env file."
+
+    return json.dumps(checks, indent=2)
+
+
 @mcp.tool()
 async def start_council(
     query: str,
@@ -426,20 +379,18 @@ async def start_council(
 ) -> str:
     """
     Phase 1: Begin a council deliberation. Runs Stage 1 (individual opinions)
-    and optionally Stage 1B (Devil's Advocate). Returns a session_id — pass it
-    to council_review() to continue.
+    and optionally Stage 1B (Devil's Advocate). 
+    
+    CRITICAL: This returns a session_id. You MUST call council_review(session_id=...) 
+    immediately after this tool to continue. DO NOT skip to council_synthesize.
     """
-    # ADR-012: Opportunistic cleanup of expired sessions
     purge_expired_sessions()
 
-    # Get configuration via lazy-loaded helpers
     tier_pools = _get_tier_model_pools()
     tier = confidence if confidence in tier_pools else "high"
     tier_contract = create_tier_contract(tier)
     tier_config = _get_tier_timeout(tier)
 
-    # Phase 1: Collect responses
-    # Note: we use our refactored run_stage1 directly
     stage1_data = await run_stage1(
         query,
         on_progress=_get_progress_callback(ctx),
@@ -448,7 +399,6 @@ async def start_council(
         adversarial_mode=adversarial_mode,
     )
 
-    # Persist session to disk
     session_id = create_session(
         query=query,
         tier=tier,
@@ -472,7 +422,7 @@ async def start_council(
 @mcp.tool()
 async def council_review(session_id: str, ctx: Context = None) -> str:
     """
-    Phase 2: Run Stage 2 peer review and ranking on a started council session.
+    Phase 2: Runs Stage 2 peer review and ranking on a started council session.
     Returns the Borda rankings table. Pass session_id to council_synthesize() next.
     """
     try:
@@ -483,7 +433,6 @@ async def council_review(session_id: str, ctx: Context = None) -> str:
     if session["stage"] != "stage1_complete":
         return json.dumps({"error": f"Session is in stage '{session['stage']}', expected 'stage1_complete'."})
 
-    # Stage 2: Peer Review
     stage2_data = await run_stage2(
         user_query=session["query"],
         stage1_data=session["stage1"],
@@ -491,18 +440,13 @@ async def council_review(session_id: str, ctx: Context = None) -> str:
         tier_contract=create_tier_contract(session["tier"]),
     )
 
-    # Fast JSON scrubbing for safety (ADR-032 / BUG-FIX)
     try:
-        # Use json.dumps with default=str to force serializability
         safe_stage2 = json.loads(json.dumps(stage2_data, default=str))
     except Exception:
-        # Fallback to crude string conversion if anything goes wrong
         safe_stage2 = {"error": "serialization_failed", "original": str(stage2_data)}
 
-    # Update session
     save_session(session_id, {"stage": "stage2_complete", "stage2": safe_stage2})
 
-    # Build preview of rankings
     aggregate = stage2_data.get("aggregate_rankings", [])
     rankings_text = "\n".join(
         f"{i+1}. {e.get('model','?').split('/')[-1]} (Borda: {float(e.get('borda_score',0)):.3f})"
@@ -524,8 +468,10 @@ async def council_synthesize(
     ctx: Context = None,
 ) -> str:
     """
-    Phase 3: Run Stage 3 Chairman synthesis on a reviewed council session.
-    Returns the full council result and cleans up the session file.
+    Phase 3: Final step — Runs Stage 3 Chairman synthesis on a reviewed council session.
+    
+    CRITICAL: This tool ONLY works after council_review() has successfully completed.
+    Returns the final synthesized verdict and cleans up the session.
     """
     try:
         session = load_session(session_id)
@@ -535,91 +481,18 @@ async def council_synthesize(
     if session["stage"] != "stage2_complete":
         return json.dumps({"error": f"Session is in stage '{session['stage']}', expected 'stage2_complete'."})
 
-    # Phase 3: Synthesis
-    # Note: we use our refactored run_stage3
     council_result = await run_stage3(
         user_query=session["query"],
         stage1_data=session["stage1"],
         stage2_data=session["stage2"],
         on_progress=_get_progress_callback(ctx),
-        verdict_type=None,  # default to synthesis
-        include_dissent=False,  # default
+        verdict_type=None,
+        include_dissent=False,
     )
 
-    # Clean up the session file
     close_session(session_id)
 
-    # Format result using our shared helper
     return _format_council_result(council_result, include_details=include_details)
-
-
-@mcp.tool()
-async def council_health_check() -> str:
-    """
-    Check LLM Council health before expensive operations (ADR-012).
-
-    Returns status of API connectivity, configured models, and estimated response time.
-    Use this to verify the council is working before calling consult_council.
-    """
-    from importlib.metadata import version as pkg_version
-
-    try:
-        council_version = pkg_version("llm-council-core")
-    except Exception:
-        council_version = "unknown"
-
-    checks = {
-        "version": council_version,
-        "api_key_configured": bool(_get_openrouter_api_key()),
-        "key_source": get_key_source(),  # ADR-013: Show where key came from (not the key itself)
-        "council_size": len(COUNCIL_MODELS),
-        "chairman_model": CHAIRMAN_MODEL,
-        "models": COUNCIL_MODELS,
-        "estimated_duration": {
-            "quick": "~20-30 seconds (fastest models)",
-            "balanced": "~45-60 seconds (most models)",
-            "high": f"~60-90 seconds (all {len(COUNCIL_MODELS)} models)",
-        },
-    }
-
-    # Quick connectivity test with a fast, cheap model
-    if checks["api_key_configured"]:
-        try:
-            start = time.time()
-            response = await query_model_with_status(
-                "google/gemini-2.0-flash-001",  # Fast and cheap
-                [{"role": "user", "content": "ping"}],
-                timeout=10.0,
-            )
-            latency_ms = int((time.time() - start) * 1000)
-
-            checks["api_connectivity"] = {
-                "status": response["status"],
-                "latency_ms": latency_ms,
-                "test_model": "google/gemini-2.0-flash-001",
-            }
-
-            if response["status"] == STATUS_OK:
-                checks["ready"] = True
-                checks["message"] = "Council is ready. Use consult_council to ask questions."
-            else:
-                checks["ready"] = False
-                checks["message"] = (
-                    f"API connectivity issue: {response.get('error', 'Unknown error')}"
-                )
-
-        except Exception as e:
-            checks["api_connectivity"] = {
-                "status": "error",
-                "error": str(e),
-            }
-            checks["ready"] = False
-            checks["message"] = f"Health check failed: {e}"
-    else:
-        checks["ready"] = False
-        checks["message"] = "OPENROUTER_API_KEY not configured. Set it in environment or .env file."
-
-    return json.dumps(checks, indent=2)
 
 
 @mcp.tool()
@@ -633,27 +506,9 @@ async def verify(
 ) -> str:
     """
     Verify agent work using the LLM Council verification system (ADR-034).
-
-    Uses multi-model consensus to verify code changes, implementations, or other
-    work artifacts against quality rubrics. Returns a structured verdict with
-    confidence score and rationale.
-
-    Args:
-        snapshot_id: Git commit SHA to verify (7-40 hex characters).
-        target_paths: Optional list of specific file paths to verify.
-        rubric_focus: Optional rubric focus area (e.g., "security", "performance").
-        confidence_threshold: Minimum confidence for pass verdict (0.0-1.0, default 0.7).
-        tier: Confidence tier for model selection - "quick", "balanced" (default), "high", or "reasoning".
-        ctx: MCP context for progress reporting (injected automatically).
-
-    Returns:
-        JSON string containing verification result with verdict, confidence,
-        exit_code (0=PASS, 1=FAIL, 2=UNCLEAR), rubric scores, blocking issues,
-        rationale, and transcript location for audit trail.
+    Uses multi-model consensus to verify code changes against quality rubrics.
     """
-
     try:
-        # Create request object and transcript store
         request = VerifyRequest(
             snapshot_id=snapshot_id,
             target_paths=target_paths,
@@ -663,41 +518,74 @@ async def verify(
         )
         store = create_transcript_store()
 
-        # Run the verification with progress callback (ADR-012 standard bridge)
         result = await run_verification(
             request, 
             store, 
             on_progress=_get_progress_callback(ctx)
         )
 
-        # Return formatted output for human readability
-        # JSON is also included at the end for programmatic parsing
         formatted = format_verification_result(result)
         json_output = json.dumps(result, indent=2)
 
         return f"{formatted}\n\n---\n\n<details>\n<summary>Raw JSON</summary>\n\n```json\n{json_output}\n```\n</details>"
 
     except InvalidSnapshotError as e:
-        return json.dumps(
-            {
-                "error": str(e),
-                "exit_code": 2,  # UNCLEAR for invalid input
-                "verdict": "unclear",
-                "confidence": 0.0,
-            },
-            indent=2,
+        return json.dumps({
+            "error": str(e), 
+            "verdict": "unclear", 
+            "confidence": 0.0,
+            "exit_code": 2
+        }, indent=2)
+    except Exception as e:
+        return json.dumps({
+            "error": f"Unexpected error: {e}", 
+            "verdict": "unclear", 
+            "confidence": 0.0,
+            "exit_code": 2
+        }, indent=2)
+
+
+
+@mcp.tool()
+async def consult_council(
+    query: str,
+    confidence: str = "balanced",
+    include_details: bool = False,
+    ctx: Context = None,
+) -> str:
+    """
+    [DEPRECATED] Run a complete monolithic council deliberation.
+    
+    This tool is maintained for backward compatibility with legacy clients.
+    For modern granular control and real-time visibility, use the sequence:
+    start_council -> council_review -> council_synthesize.
+    
+    Args:
+        query: Your question for the council
+        confidence: quick, balanced, or high (default: balanced)
+        include_details: If true, includes individual model responses and rankings
+    """
+    # Check if we are in a test environment (ADR-012 fix)
+    import os, sys
+    in_test = os.environ.get("PYTEST_CURRENT_TEST") or "pytest" in sys.modules
+    
+    if ctx is not None and not in_test:
+        return (
+            "ERROR: consult_council is disabled for real-time use to prevent timeouts. "
+            "Please use the new granular tools: start_council -> council_review -> council_synthesize."
         )
 
-    except Exception as e:
-        return json.dumps(
-            {
-                "error": f"Unexpected error: {e}",
-                "exit_code": 2,  # UNCLEAR for unexpected errors
-                "verdict": "unclear",
-                "confidence": 0.0,
-            },
-            indent=2,
-        )
+    # Use run_council_with_fallback for monolithic execution
+    # This maintains the ADR-012 structured response format
+    on_progress = _get_progress_callback(ctx)
+    result = await run_council_with_fallback(
+        query,
+        on_progress=on_progress,
+        tier_contract=create_tier_contract(confidence),
+        include_dissent=True,
+    )
+    
+    return _format_council_result(result, include_details=include_details)
 
 
 @mcp.tool()
@@ -709,83 +597,66 @@ async def audit(
 ) -> str:
     """
     Retrieve and validate verification audit transcripts (ADR-034).
-
-    Provides access to verification audit trails for compliance, debugging,
-    and integrity validation. Can retrieve a single verification by ID or
-    list all verifications.
-
-    Args:
-        verification_id: Optional ID to retrieve specific verification.
-            If not provided, lists all available verifications.
-        validate_integrity: If True, validates transcript integrity against
-            expected_hash.
-        expected_hash: Expected SHA256 hash for integrity validation.
-            Required when validate_integrity is True.
-        ctx: MCP context (injected automatically).
-
-    Returns:
-        JSON string containing:
-        - For single verification: stages, integrity_hash, optional validation result
-        - For listing: verifications array with metadata, total_count
+    Provides access to verification audit trails for compliance and debugging.
     """
     try:
         store = create_transcript_store(readonly=True)
 
-        # If no verification_id, list all verifications
         if verification_id is None:
             verifications = store.list_verifications()
-            return json.dumps(
-                {
-                    "verifications": verifications,
-                    "total_count": len(verifications),
-                },
-                indent=2,
-            )
+            return json.dumps({"verifications": verifications, "total_count": len(verifications)}, indent=2)
 
-        # Retrieve specific verification
-        stages = store.read_all_stages(verification_id)
-        integrity_hash = store.compute_integrity_hash(verification_id)
+        try:
+            store = create_transcript_store(verification_id)
+            # Compatibility: Support both old and new method names
+            if hasattr(store, "read_all_stages"):
+                transcript = store.read_all_stages()
+            else:
+                transcript = store.get_transcript()
 
-        result: dict = {
-            "verification_id": verification_id,
-            "stages": stages,
-            "integrity_hash": integrity_hash,
-        }
+            if not transcript:
+                return json.dumps({"error": f"Verification {verification_id} not found"}, indent=2)
 
-        # Validate integrity if requested
-        if validate_integrity and expected_hash:
-            try:
-                store.validate_integrity(verification_id, expected_hash)
-                result["integrity_valid"] = True
-            except TranscriptIntegrityError as e:
-                result["integrity_valid"] = False
-                result["integrity_error"] = str(e)
+            # Compatibility: Support both old and new hash methods
+            if hasattr(store, "compute_integrity_hash"):
+                actual_hash = store.compute_integrity_hash()
+            else:
+                actual_hash = store.get_transcript_hash()
 
-        return json.dumps(result, indent=2)
-
-    except TranscriptNotFoundError as e:
-        return json.dumps(
-            {
-                "error": f"Verification not found: {e}",
+            result = {
                 "verification_id": verification_id,
-            },
-            indent=2,
-        )
+                "transcript": transcript,
+                "stages": transcript,  # Backward compatibility for tests
+                "integrity_hash": actual_hash,  # Backward compatibility
+                "integrity_validation": {"valid": True, "hash": actual_hash}
+            }
 
+            if validate_integrity:
+                try:
+                    store.validate_integrity(expected_hash)
+                    result["integrity_valid"] = True
+                except Exception as e:
+                    result["integrity_valid"] = False
+                    result["integrity_error"] = str(e)
+                    result["integrity_validation"]["valid"] = False
+                    result["integrity_validation"]["error"] = str(e)
+
+            return json.dumps(result, indent=2)
+        except Exception as e:
+            return json.dumps({
+                "error": str(e),
+                "exit_code": 2,  # UNCLEAR
+                "verdict": "error"
+            }, indent=2)
     except Exception as e:
-        return json.dumps(
-            {
-                "error": f"Unexpected error: {e}",
-                "verification_id": verification_id,
-            },
-            indent=2,
-        )
+        return json.dumps({"error": str(e)}, indent=2)
 
+
+# --- Entry Point ---
 
 def main():
     """Entry point for the llm-council command."""
     mcp.run()
-
 
 if __name__ == "__main__":
     main()
